Added SensorService for Accl, Gyro, Mag, Lux sensors | PR Sensors (#1)

+++ lib/services/models/sensor_models.dart
@@ -17,7 +17,7 @@ class ImuSensorData {
 }
 
 class LuxSensorData {
-  final double lux;
+  final int lux;
   final DateTime timeStamp;
   final SensorState state;
